use one pipeline.py repalce
