--- dtoolkit/pipeline/__init__.py
@@ -1,4 +1,0 @@
-from dtoolkit.pipeline.FeatureUnion import FeatureUnion  # noqa
-from dtoolkit.pipeline.make_pipeline import make_pipeline  # noqa
-from dtoolkit.pipeline.make_union import make_union  # noqa
-from dtoolkit.pipeline.Pipeline import Pipeline  # noqa
