--- dtoolkit/pipeline/make_union.py
@@ -1,41 +1,0 @@
-from __future__ import annotations
-
-from sklearn.pipeline import _name_estimators
-
-from dtoolkit.pipeline.FeatureUnion import FeatureUnion
-from dtoolkit.transformer import Transformer
-
-
-def make_union(
-    *transformers: list[Transformer],
-    n_jobs: int = None,
-    verbose: bool = False,
-) -> FeatureUnion:
-    """
-    Construct a FeatureUnion from the given transformers.
-
-    See Also
-    --------
-    FeatureUnion
-        Class for concatenating the results of multiple transformer objects.
-
-    Notes
-    -----
-    Different to :obj:`sklearn.pipeline.make_union`.
-    This would let :obj:`~pandas.DataFrame` in and
-    :obj:`~pandas.DataFrame` out.
-
-    Examples
-    --------
-    >>> from sklearn.decomposition import PCA, TruncatedSVD
-    >>> from dtoolkit.pipeline import make_union
-    >>> make_union(PCA(), TruncatedSVD())
-     FeatureUnion(transformer_list=[('pca', PCA()),
-                                   ('truncatedsvd', TruncatedSVD())])
-    """
-
-    return FeatureUnion(
-        _name_estimators(transformers),
-        n_jobs=n_jobs,
-        verbose=verbose,
-    )
